+++ pkg/sql/opt/norm/prune_cols_funcs.go
@@ -631,3 +631,21 @@ func (c *CustomFuncs) PruneMutationReturnCols(
 	newPrivate.PassthroughCols = newPassthroughCols
 	return &newPrivate
 }
+
+// MutationTable returns the table upon which the mutation is applied.
+// CHECK
+func (c *CustomFuncs) MutationTable(private *memo.MutationPrivate) opt.TableID {
+	return private.Table
+}
+
+// NeededColMapLeft returns the subset of a SetPrivate's LeftCols that corresponds to the
+// needed subset of OutCols. This is useful for pruning columns in set operations.
+func (c *CustomFuncs) NeededColMapLeft(needed opt.ColSet, set *memo.SetPrivate) opt.ColSet {
+	return opt.TranslateColSet(needed, set.OutCols, set.LeftCols)
+}
+
+// NeededColMapRight returns the subset of a SetPrivate's RightCols that corresponds to the
+// needed subset of OutCols. This is useful for pruning columns in set operations.
+func (c *CustomFuncs) NeededColMapRight(needed opt.ColSet, set *memo.SetPrivate) opt.ColSet {
+	return opt.TranslateColSet(needed, set.OutCols, set.RightCols)
+}
